new animation system!
In the old animation system, each animation "stack" would have a movement position and an animation type, requiring animations and movements to happen together. It was also pretty cumbersome.
With the new implementation, animations and movements are seperate, requiring less boolean juggling to keep things in check and also creating/destroying tween nodes as needed to get rid of the weirdness I was getting from just using one!